fix: tighten dba_vector_analysis verify, inline vectors_setup, defensive role cleanup in rls_business_access
